+++ src/cmd.rs
@@ -254,7 +254,6 @@ fn connect_(
             addr: serv_name.to_owned(),
             port: serv_port,
             tls: defaults.tls,
-            hostname: defaults.hostname.clone(),
             realname: defaults.realname.clone(),
             pass: pass.map(str::to_owned),
             nicks: defaults.nicks.clone(),